remove usage of the recoverability property
