+++ src/NServiceBus.AcceptanceTesting/Support/EndpointRunner.cs
@@ -6,6 +6,7 @@
     using System.Threading;
     using System.Threading.Tasks;
     using Configuration.AdvancedExtensibility;
+    using Faults;
     using Logging;
     using NServiceBus.Support;
     using Transport;
@@ -54,6 +55,27 @@
                 }
                 endpointConfiguration = await configuration.GetConfiguration(run).ConfigureAwait(false);
                 RegisterInheritanceHierarchyOfContextInSettings(scenarioContext);
+                endpointConfiguration.Recoverability().Failed(settings => settings.OnMessageSentToErrorQueue(m =>
+                {
+                    scenarioContext.FailedMessages.AddOrUpdate(
+                        endpointName,
+                        new[]
+                        {
+                            new FailedMessage(m.MessageId, new Dictionary<string, string>(m.Headers), m.Body, m.Exception, m.ErrorQueue)
+                        },
+                        (i, failed) =>
+                        {
+                            var result = failed.ToList();
+                            result.Add(new FailedMessage(m.MessageId, new Dictionary<string, string>(m.Headers), m.Body, m.Exception, m.ErrorQueue));
+                            return result;
+                        });
+
+                    //We need to set the error flag to false as we want to reset all processing exceptions caused by immediate retries
+                    scenarioContext.UnfinishedFailedMessages.AddOrUpdate(m.MessageId, id => false, (id, value) => false);
+
+                    return Task.FromResult(0);
+                }));
+                endpointConfiguration.Pipeline.Register(new CaptureExceptionBehavior(scenarioContext.UnfinishedFailedMessages), "Captures unhandled exceptions from processed messages for the AcceptanceTesting Framework");
 
                 endpointBehavior.CustomConfig.ForEach(customAction => customAction(endpointConfiguration, scenarioContext));
